Updated documentation

+++ src/word.h
@@ -1,7 +1,7 @@
 struct se_word {
   unsigned char type;
   unsigned char refs;
-  int data;
+  long int data;
 };
 
 struct se_word se_add(const struct se_word a, const struct se_word b) {
